+++ pkg/queue/queue.go
@@ -1,9 +1,9 @@
 package queue
 
 type Queue interface {
-	Enqueue(priority int, content string) (*Message, error)
+	Enqueue(priority int64, content string) (*Message, error)
 	Dequeue() (*Message, error)
 	GetByID(id uint64) (*Message, error)
-	UpdatePriority(id uint64, newPriority int) error
+	UpdatePriority(id uint64, newPriority int64) error
 	Len() int
 }
